feat(S4): 랜딩 페이지 로직 수정
